requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.18.1
idna==3.3
Kivy-Garden==0.1.5
Pillow==9.1.0
pycodestyle==2.8.0
Pygments==2.11.2
requests==2.27.1
tk==0.1.0
toml==0.10.2
urllib3==1.26.9
